Add BinFheScheme weaver, cone-fusion plan builder, and plan codegen (M6)
volar-weaver gains fhe_binfhe: an FheScheme impl over the V2 gate
surface (all gates = single-bootstrap LUT reads), a cone-fusion pass
that collapses maximal And/Or/Xor/Not cones into multi-input LUT ops
(one blind rotation per cone) with arity bounds and a checked failure
budget, and weave_binfhe_plan emitting an IrModule from a validated
BootstrapPlan. Generated code calls binfhe_lut_read_dyn — the same
executor as the spec's execute_plan — and the e2e test proves the
generated function runs bit-identically to the interpreter on the toy
profile. Spec side: gates take wires by value (legacy parity),
k_max is usize at fn surfaces, and execute_plan shares the dyn LUT
executor.

+++ crates/spec/volar-spec/src/binfhe/circuit_bs.rs
@@ -223,7 +223,7 @@ fn priv_ks<const BIG_N: usize, const LOG_Q: u32, const PRIV_ELL: usize, const PR
 fn level_test_poly<const BIG_N: usize, const LOG_Q: u32>(
     level: usize,
     bs_base_log: u32,
-    k_max: u32,
+    k_max: usize,
 ) -> [u32; BIG_N] {
     let width = BIG_N >> k_max;
     let g = gadget::level_factor::<LOG_Q>(bs_base_log, level);
@@ -250,9 +250,9 @@ pub fn circuit_bootstrap<
 >(
     ct: &LweCiphertext<N_LWE>,
     cbk: &CircuitBootstrappingKey<N_LWE, BIG_N, BS_ELL, KS_ELL, PRIV_ELL>,
-    k_max: u32,
+    k_max: usize,
 ) -> RgswCiphertext<BIG_N, BS_ELL> {
-    let delta = wire_delta::<LOG_Q_LWE>(k_max);
+    let delta = wire_delta::<LOG_Q_LWE>(k_max as usize);
     // Center the wire's bin: phase becomes m * Delta + Delta/2.
     let centered = lwe_add_const::<N_LWE, LOG_Q_LWE>(ct, delta / 2);
     let rows = core::array::from_fn(|j| {
@@ -285,7 +285,7 @@ mod tests {
     use crate::binfhe::lwe::{gen_lwe_secret_key, lwe_encrypt};
     use crate::binfhe::params::toy;
     use crate::binfhe::rgsw::{cmux, external_product};
-    use crate::binfhe::rlwe::{gen_rlwe_secret_key, rlwe_phase, rlwe_trivial};
+    use crate::binfhe::rlwe::{gen_rlwe_secret_key, rlwe_trivial};
 
     struct TestRng(u64);
     impl TestRng {
